Client uses SSO to set cookie
When the client app first loads it makes a request to retrieve its SSO
session if there is one.  If there is, it uses this to populate the
account menu.  If not, it shows a "login" button which shows a form in a
modal dialog.

The form makes an XHR post to the SSO server making sure that any
cookies in the response are set.

All cookies set by the SSO server are `HttpOnly`, `Same-Site=Strict` and
`Secure`.