Prevent `relative_path` breaking for some jobs
Getting the `relative_path` for stdout and stderr assumes that a job has
reported its `results_dir`.  This will not be the case for old jobs.
The client isn't using the relative path for stdout or stderr so the
quick fix is to return `nil` for those files.  This should be fixed
properly later.