Simplify thread reading issues
Previously, we were trying to handle a case where the child process
had completed without closing its stdout and/or stderr process.  This
might be possible if the child process creates child processes that end
up being zombied.

The concern was that if the stdout pipe had not been closed, we couldn't
trust that we had read all of its output and hence couldn't trust what
we had read.

Now, we use a simpler mechanism for determining whether we can trust the
output.  For commands that are expected to return a JSON response, we
trust the output if it is valid JSON.  For other commands we simply
trust the output.

This is a much simpler, more general method of handling a very unlikely
edge case.