Fix clearing cache
`useFetch` has a couple of different ways of being called.

1. Pass an array of dependencies as the third argument to `useFetch`.
   This will cause the request to be made immediately and again each
   time the given array of dependencies changes.  This is how we make
   GET requests.

2. Do not pass a third arguments to `useFetch`.  The request will not be
   executed until `patch`, `put`, `del`, `get` or `post` is called on
   the request object.  This is how we make `POST` requests.

The important thing to not above is that we cannot clear the cache when
we construct the request in the second form, we need to wait until the
request is executed.

I've added an response interceptor which clears the cache when the
relevant requests are exectued.  It only does so if the response is OK.

I've also changed the way in which the cache was being cleared.  We now
clear the cache via the `useFetch` API.  We can get a copy of the entire
cache by calling `useFetch` without a path/url.

I've also made sure that we clear both the memory and persisted caches.
Probably overkill.